Allow configuration variables to be set during deployment
Here we add a new optional parameter to the action so that users
may set aptible config variables during the deploy action. This
is useful for cases where config values might depend on information
present during the Github workflow that is calling the action
(for example, the git sha used to build the image).

Example usage is added to the documentation as well.